+++ assets/js/modules/sign-in-with-google/components/types.ts
@@ -1,0 +1,30 @@
+/**
+ * Sign in with Google shared component types.
+ *
+ * Site Kit by Google, Copyright 2026 Google LLC
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     https://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+export type PluginWithConflict = {
+	pluginName: string;
+	conflictMessage?: string | null;
+};
+
+export type SignInWithGoogleCompatibilityErrors = {
+	/* eslint-disable camelcase */
+	conflicting_plugins?: { [ slug: string ]: PluginWithConflict };
+	wp_login_inaccessible?: boolean;
+	host_wordpress_dot_com?: boolean;
+	/* eslint-enable camelcase */
+};
